backend/requirements.txt
# 基础依赖
python-dotenv

# FastAPI 框架相关
uvicorn
fastapi==0.123.0
pydantic
aiofiles
python-multipart
python-jose[cryptography]

# LangChain 核心库
langchain
langchain-core
langchain-community
langchain-chroma
pypdf

# 向量数据库
chromadb

# 数据库相关
sqlalchemy
aiomysql